requirements.txt
# Core dependencies
pyyaml >= 6.0
click >= 8.0

# Development dependencies
pytest >= 7.0
pytest-cov >= 4.0
colorama >= 0.4.4